Add support for value type arrays
This required a few improvements to language-cil, so for the time being
we're back at depending on bamboo/language-cil@idris-cil.

In the process I had to remove `CIL.FFI.Array.get` and replace it with
`CIL.FFI.Array.get'` which requires the array element type as the first
argument. I couldn't get the element type index stored in TypedArray
to flow all the way to the code generator. The search for a better
solution continues.

See #5